improving invitations

+++ resources/views/components/card.blade.php
@@ -1,10 +1,12 @@
-<div class="card h-48 flex flex-col">
+<div class="card flex flex-col {{ $height ?? '' }}">
     <div class="text-xl font-normal py-4 -ml-5 border-l-4 border-blue-custom pl-4 mb-3">
         <a href="{{ route('projects.show', $project) }}">{{ $project->title }}</a>
     </div>
     <p class="text-gray-dark mb-4 flex-1">{{ Str::limit($project->description, 150) }}</p>
 
     <footer class="text-right">
-        <a href="{{ route('projects.destroy', $project) }}" data-confirm="Are you sure?" data-method="delete" rel="nofollow" class="button-sm-red">{{ __('Delete') }}</a>
+        @can('delete', $project)
+            <a href="{{ route('projects.destroy', $project) }}" data-confirm="Are you sure?" data-method="delete" rel="nofollow" class="button-sm-red">{{ __('Delete') }}</a>
+        @endcan
     </footer>
 </div>
